delhi-ps-crm-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
supabase==2.4.6
python-dotenv==1.0.1
google-genai==1.73.1
apscheduler==3.10.4
joblib==1.4.2
scikit-learn==1.5.2
pydantic==2.7.1
numpy==2.3.3
